+++ src/edlib.cpp
@@ -407,7 +407,7 @@ static inline std::vector<int> getBlockCellValues (const Block& block) {
     int score = block.score;
     Word mask = HIGH_BIT_MASK;
     for (int i = 0; i < WORD_SIZE - 1; i++) {
-        scores[i] = score;
+        scores[static_cast<unsigned long>(i)] = score;
         if (block.P & mask) {
             score--;
         }
@@ -472,7 +472,7 @@ static inline void readBlockReverse (const Block& block, int* const dest) {
 static inline bool allBlockCellsLarger (const Block& block, const int k) {
     std::vector<int> scores = getBlockCellValues(block);
     for (int i = 0; i < WORD_SIZE; i++) {
-        if (scores[i] <= k) {
+        if (scores[static_cast<unsigned long>(i)] <= k) {
             return false;
         }
     }
@@ -642,7 +642,7 @@ static int myersCalcEditDistanceSemiGlobal (
     if (lastBlock == maxNumBlocks - 1) {
         std::vector<int> blockScores = getBlockCellValues(*bl);
         for (int i = 0; i < W; i++) {
-            int colScore = blockScores[i + 1];
+            int colScore = blockScores[static_cast<unsigned long>(i + 1)];
             if (colScore <= k && (bestScore == -1 || colScore <= bestScore)) {
                 if (colScore != bestScore) {
                     positions.clear();
@@ -812,7 +812,8 @@ static int myersCalcEditDistanceNW (const Word* const Peq, const int W, const in
                 bool reduce = true;
                 for (int i = WORD_SIZE - numCells; i < WORD_SIZE; i++) {
                     // TODO: Does not work if do not put +1! Why???
-                    if (scores[i] <= k && r <= k - scores[i] - targetLength + c + queryLength + 1) {
+                    if (scores[static_cast<unsigned long>(i)] <= k &&
+                            r <= k - scores[static_cast<unsigned long>(i)] - targetLength + c + queryLength + 1) {
                         reduce = false;
                         break;
                     }
@@ -832,7 +833,8 @@ static int myersCalcEditDistanceNW (const Word* const Peq, const int W, const in
                 int r = firstBlock * WORD_SIZE + numCells - 1;
                 bool reduce = true;
                 for (int i = WORD_SIZE - numCells; i < WORD_SIZE; i++) {
-                    if (scores[i] <= k && r >= scores[i] - k - targetLength + c + queryLength) {
+                    if (scores[static_cast<unsigned long>(i)] <= k &&
+                            r >= scores[static_cast<unsigned long>(i)] - k - targetLength + c + queryLength) {
                         reduce = false;
                         break;
                     }
@@ -889,7 +891,7 @@ static int myersCalcEditDistanceNW (const Word* const Peq, const int W, const in
 
     if (lastBlock == maxNumBlocks - 1) { // If last block of last column was calculated
         // Obtain best score from block -> it is complicated because query is padded with W cells
-        int bestScore = getBlockCellValues(blocks[lastBlock])[W];
+        int bestScore = getBlockCellValues(blocks[lastBlock])[static_cast<unsigned long>(W)];
         if (bestScore <= k) {
             *bestScore_ = bestScore;
             *position_ = targetLength - 1;
